fix: span propagation across microservices
